// src/lib/types/Geometry.ts
export type Geometry = {
  frameSize: number;
  seatTubeLength: number;
  topTubeHorizontal: number;
  topTubeActual?: number;
  headTubeAngle: number;
  seatTubeAngle: number;
  standover: number;
  headTubeLength: number;
  wheelbase: number;
  frontCenter?: number;
  chainStayLength: number;
  bottomBracketDrop: number;
  bottomBracketHeight?: number;
  forkRake?: number;
  trail?: number;
  stack: number;
  reach: number;
};